refactor : 배럴 정리
