+++ src/features/professor-search/index.ts
@@ -1,0 +1,2 @@
+export { ProfessorResultList } from "@/features/professor-search/ui/ProfessorResultList";
+export { SearchProfessorToCompare } from "@/features/professor-search/ui/SearchProfessorToCompare";
