+++ src/features/university-compare/index.ts
@@ -1,0 +1,1 @@
+export { UniversityCard } from "@/features/university-compare/ui/UniversityCard";
