+++ src/features/university-search/index.ts
@@ -1,0 +1,2 @@
+export { SearchUniversityToCompare } from "@/features/university-search/ui/SearchUniversityToCompare";
+export { UniversityResultList } from "@/features/university-search/ui/UniversityResultList";
